src/ez-postgres/resources/requirements.txt
boto3==1.21.36
botocore==1.24.36
jmespath==1.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
